// playwright-e2e/constants/ccd-events/create-claim/personal-injury-type.ts
enum PersonalInjuryType {
  ROAD_ACCIDENT = 'ROAD_ACCIDENT',
  WORK_ACCIDENT = 'WORK_ACCIDENT',
  PUBLIC_LIABILITY = 'PUBLIC_LIABILITY',
  HOLIDAY_ILLNESS = 'HOLIDAY_ILLNESS',
  DISEASE_CLAIM = 'DISEASE_CLAIM',
  NOISE_INDUCED_HEARING_LOSS = 'NOISE_INDUCED_HEARING_LOSS',
  PERSONAL_INJURY_OTHER = 'PERSONAL_INJURY_OTHER',
}

export default PersonalInjuryType;