// playwright-e2e/constants/cases/claim-type.ts
enum ClaimType {
  ONE_VS_ONE_LIP = '1v1LIP',
  TWO_VS_ONE_LIP = '2v1LIP',
  ONE_VS_TWO_LIPS = '1v2LIPS',
  ONE_VS_TWO_LR_LIP = '1v2LRLIP',
  ONE_VS_TWO_LIP_LR = '1v2LIPLR',
  ONE_VS_ONE = '1v1',
  TWO_VS_ONE = '2v1',
  ONE_VS_TWO_SAME_SOL = '1v2SS',
  ONE_VS_TWO_DIFF_SOL = '1v2DS',
}

export default ClaimType;